move updating alerts logic to root
